backend/requirements.txt
flask==2.2.3
flask-socketio==5.3.2
flask-cors==3.0.10
eventlet==0.33.3
werkzeug==2.2.3
google-generativeai